requirements.txt
# para instalar las librerias necesarias, usar:
# pip install -r requirements.txt
numpy
Pillow
rasterio
amulet-core 
amulet-nbt
dotenv
scikit-learn
scipy